+++ ChangeLog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ### 1.5.0 (29 May 2016)
- * Fix: Tries to address an incompatibility between WordPress and PHP Console: "PHP Warning: session_start(): Cannot send session cache limiter - headers already sent"
+ * Fix: Fixes "PHP Warning: session_start(): Cannot send session cache limiter - headers already sent" notice in logs
  * Misc: Internal changes, new Settings class, deprecated methods and properties in main Plugin class
  * Misc: Updated PHP Console Library to 3.1.6
  * Misc: Tested up to WordPress 4.5.2